requirements.txt
anthropic
requests
python-dotenv
splunk-sdk
